feat: 마이페이지 cId 세션 하나로 접속하기

+++ src/main/java/com/example/greenspringboot/cust/entity/Cust.java
@@ -95,8 +95,7 @@ public class Cust {
 
     @Builder.Default
     @Column(name= "visit_cnt")
-    private Integer visitCnt = 0;
-
+    private long visitCnt = 0;
 
     @Builder.Default
     private Date frst_reg_dt = new Date();
